chore: moved files to proper working directory
